Almost Done!
- API endpoints added for creation, elimination and updates
- Front-end started and almost finished

Pending:
- Angular addition
- Endpoint for getting tasks

+++ Backend/DB_Controller.cs
@@ -1,10 +1,13 @@
 using MongoDB.Bson;
 using MongoDB.Driver;
+using Microsoft.OpenApi.Exceptions;
+using Microsoft.OpenApi.Validations;
 
 
 namespace DBControl {
     class CRUD_Control {
     public void CreateTask(string Name, string Details) {
+        var Success = (string Mg) => Mg;
         var Client = new MongoClient("mongodb://localhost:27017");
         var DataB = Client.GetDatabase("D-NET-DB");
         var Collection = DataB.GetCollection<BsonDocument>("Tasks");
@@ -12,15 +15,22 @@ namespace DBControl {
             {"Task_Name", Name},
             {"Task_Details", Details}
         };
-        Collection.InsertOne(Task);
+        var filter = Builders<BsonDocument>.Filter.Eq("Task_Name", Name);
+        var Check = Collection.Find(filter);
+        var Check_Count = Check.CountDocuments();
+        int CorrectCount = (int) Check_Count;
+        if(CorrectCount >= 1) {
+            throw new InvalidOperationException("There is already a task with this name");
+        } else {
+            Collection.InsertOne(Task);
+        }
     }
 
-    public void GetTasksU() {
-
+    public void GetTasks() {
         var Client = new MongoClient("mongodb://localhost:27017");
         var DataB = Client.GetDatabase("D-NET-DB");
         var Collection = DataB.GetCollection<BsonDocument>("Tasks");
-        var DocumentList = Collection.Find(new BsonDocument()).ToList();
+        var DocumentList = Collection.Find(new BsonDocument()).ToListAsync();
         System.Console.WriteLine(DocumentList);
 
     }
@@ -31,7 +41,12 @@ namespace DBControl {
         var Collection = DataB.GetCollection<BsonDocument>("Tasks");
         var Filter = Builders<BsonDocument>.Filter.Eq("Task_Name", OldName);
         var N_Update = Builders<BsonDocument>.Update.Set("Task_Name", NewName);
-        Collection.UpdateOne(Filter, N_Update);
+        var Query = Collection.Find(Filter).CountDocuments();
+        if(Query == 0) {
+            throw new InvalidOperationException("This task does not exist");
+        } else {
+            Collection.UpdateOne(Filter, N_Update);
+        }
     }
 
     public void UpdateTaskForDetails(string OldName, string NewDetails) {
@@ -40,7 +55,12 @@ namespace DBControl {
         var Collection = DataB.GetCollection<BsonDocument>("Tasks");
         var Filter = Builders<BsonDocument>.Filter.Eq("Task_Name", OldName);
         var D_Update = Builders<BsonDocument>.Update.Set("Task_Details", NewDetails);
-        Collection.UpdateOne(Filter, D_Update)
+        var Query = Collection.Find(Filter).CountDocuments();
+        if(Query == 0) {
+            throw new InvalidOperationException("This task does not exist");
+        } else {
+            Collection.UpdateOne(Filter, D_Update);
+        }
     }
 
     public void DeleteTask(string OldName) {
@@ -48,7 +68,12 @@ namespace DBControl {
         var DataB = Client.GetDatabase("D-NET-DB");
         var Collection = DataB.GetCollection<BsonDocument>("Tasks");
         var Filter = Builders<BsonDocument>.Filter.Eq("Task_Name", OldName);
-        Collection.DeleteOne(Filter);
+        var Query = Collection.Find(Filter).CountDocuments();
+        if(Query == 0) {
+            throw new InvalidOperationException("This task does not exist");
+        } else {
+            Collection.DeleteOne(Filter);
+        }
     }
 };
 }
